+++ sdks/go/pkg/beam/io/filesystem/gcs/gcs.go
